Function cat_data_service::service::api::health::retry_after_put::endpoint

pub(crate) async fn endpoint(
    params: Result<Query<RetryAfterParams>>,
    state: Data<&Arc<State>>
) -> UniResponse<Empty, Empty, Empty, Empty, Empty, Empty, Empty, NoContent, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ApiValidationError, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ServerError, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty>

PUT /health/retry-after

Configuration of the “retry-after” for 503 responses endpoint.

During the service runtime, when disconnects to the database happen, the appropriate response status code is 503, which includes a string that is either <delay-secods> or <http-date>.

This endpoint accepts two optional query parameters:

http-date is the UTC Datetime string, which if set, will be returned in all 503 responses. Example: PUT /health/retry-after?http-date=2099-01-31T00:00:00Z

delay-seconds is number of seconds as a positive integer, which if set, will be returned in all 503 response. Example: PUT /health/retry-after?delay-seconds=120

If none of the query parameters is present, then the header value will be reset to the service’s original settings. Example: PUT /health/retry-after

If no problems occur, this endpoint will always just return 204.

If query parameters are invalid, for example an invalid datetime strinn In theory it can also return 500 if for some reason a temporary circumstance is preventing this service from properly serving request.

An example could be the service has started a long and cpu intensive task and is not able to properly service requests while it is occurring. This would let the load balancer shift traffic to other instances of this service that are ready.

Responses

204 No Content - Service is Ready to serve requests.

400 API Validation Error - If anything within this function fails unexpectedly. (Possible but unlikely)

500 Service - Service is not ready, do not send other requests.