Module cat_data_service::settings

Command line and environment variable settings for the service

Structs

API_HOSTNAMES 🔒

A List of servers to provideThe client id key used to anonymize client connections.

API_URL_PREFIX 🔒

The Basepath the API is served at.

CLIENT_ID_KEY 🔒

The client id key used to anonymize client connections.

GITHUB_ISSUE_TEMPLATE 🔒

The github issue template to use

GITHUB_REPO_NAME 🔒

The github repo name

GITHUB_REPO_OWNER 🔒

The github repo owner

RetryAfterParams 🔒

Settings

StringEnvVar 🔒

An environment variable read as a string.

Constants

ADDRESS_DEFAULT 🔒

Default address to start service on.

API_HOSTNAMES_DEFAULT 🔒

Default API_HOSTNAME/S used in production. This can be a single hostname, or a list of them.

API_URL_PREFIX_DEFAULT 🔒

Default API_URL_PREFIX used in development.

CLIENT_ID_KEY_DEFAULT 🔒

Default CLIENT_ID_KEY used in development.

GITHUB_ISSUE_TEMPLATE_DEFAULT 🔒

Default Github issue template to use

GITHUB_REPO_NAME_DEFAULT 🔒

Default Github repo name

GITHUB_REPO_OWNER_DEFAULT 🔒

Default Github repo owner

RETRY_AFTER_DELAY_SECONDS_DEFAULT 🔒

The default delay-seconds in the ‘Retry-After’ header of 503 responses.

RETRY_AFTER_DELAY_SECONDS_ENVVAR 🔒

The name of the env var that specifies the delay-seconds in the ‘Retry-After’ header of 503 responses.

RETRY_AFTER_HTTP_DATE_DEFAULT 🔒

The default http-date in the ‘Retry-After’ header of 503 responses.

RETRY_AFTER_HTTP_DATE_ENVVAR 🔒

The name of the env var that specifies the delay-seconds in the ‘Retry-After’ header of 503 responses.

Statics

LAZY 🔒

LAZY 🔒

LAZY 🔒

LAZY 🔒

LAZY 🔒

LAZY 🔒

Functions

__stability 🔒

__stability 🔒

__stability 🔒

__stability 🔒

__stability 🔒

__stability 🔒

__static_ref_initialize 🔒

__static_ref_initialize 🔒

__static_ref_initialize 🔒

__static_ref_initialize 🔒

__static_ref_initialize 🔒

__static_ref_initialize 🔒

generate_github_issue_url 🔒

Generate a github issue url with a given title

get_api_hostnames 🔒

Get a list of all hostnames to serve the API on.

invalid_hostname 🔒

string_to_api_hostnames 🔒

Transform a string list of hostnames into a vec of hostnames. Default to the service address if none specified.